Moved `logFetchMoreResults` to new file
https://coveord.atlassian.net/browse/KIT-170
